pyproject.toml
[project]
name = "haunt"
version = "0.1.0"
description = "Python web-based framework"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pywebview>=5.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
